outsourced the IO code and removed redundant example code from main
